R6 Class Representing a `crcexperiment`

This class implements a `crcexperiment`.

Public fields

models

is a list containing crcmodel objects.

experimental_design

is a table containing one row per experiment to be ran.

grid

is a table containing one row per point in the grid experimental design.

lhs

is a table containing one row per point in the lhs experimental design.

posteriors

is a table containing one row per parameter set defined in the posterior of each model included in the experiment.

experimental_parameters

is a list containing details about each experimental parameter. Experimental parameters can be either policy levers or uncertainties. Defining this distinction is up to the user.

Methods

Public methods

crcexperiment$new()

crcexperiment$set_parameter()

crcexperiment$set_design()

crcexperiment$clone()

Method new()

Create a new `crcexperiment` object.

Usage

crcexperiment$new(...)

Arguments

...

set of crcmodels to be included in the experiment.

Returns

s new `crcexperiment` object.

Method set_parameter()

Set Parameter

Usage

crcexperiment$set_parameter(
  parameter_name,
  experimental_design,
  values,
  min,
  max
)

Arguments

parameter_name

character string defining the parameter name.

experimental_design

Either "grid" or "lhs" Use lhs if you want to create a Latin Hypercube Sample within the min and max bounds you provided. Use Grid

values

use when experimental_design = "grid". This should be a vector including the values to be included in a grid experimental design. Please use parameters and values that can be converted to strings without any issues.

min

use when experimental_design = "lhs". This should bea numeric value indicating the minimum bound in the Latin Hypercube sample.

max

use when experimental_design = "lhs". This should bea numeric value indicating the minimum bound in the Latin Hypercube sample.

Details

This function constructs the experimental_parameter object, and appends experimental parameters that will be visible inside the model in the future. Experimental parameters can be either uncertainties or decision levers. Every parameter defined in this function can be accessed within the model by using experimental_parameters$param_name.

Method set_design()

Set Experimental Design

Usage

crcexperiment$set_design(
  n_lhs,
  convert_lhs_to_grid = F,
  lhs_to_grid_midpoints = 0
)

Arguments

n_lhs

The number of points in the Latin Hypercube Sample to be created.

convert_lhs_to_grid

Default is FALSE. If TRUE, this function convert the LHS parameters to a grid design. Often is used when testing the "corners" of the experimental design before performing a full LHS run.

lhs_to_grid_midpoints

Only relevant when convert_to_lhs = T. Default value is 0. This should be an integer determining how many points within the grid hypercube should be created for the parameters being converted from LHS to a GRID design. For example, if convert_lhs_to_grid = T and lhs_to_grid_midpoints = 0, this function will create a full factorial design of the LHS parameters with 2^n points. If one wants to use one midpoint, then the design will have 3^n points, and so on. This parameter does not affect parameters orignally defined as part of a grid design because their values have already been set.

Details

Creates the future_experimental_design data.frame based on the paramers defined by the set_parameter functions. The experimental design created by this function is useful to run a typical RDM analysis where each policy is evaluated across a LHS of deep uncertainties. To achieve that, define each policy lever as a grid parameter, and each uncertainty as an "lhs" uncertainty.

Method clone()

The objects of this class are cloneable with this method.

Usage

crcexperiment$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.